+++ src/headless/plugins/reactions/index.js
@@ -1,0 +1,5 @@
+/**
+ * XEP-0444: Message Reactions - Headless core
+ */
+
+import './plugin.js';