+++ src/headless/shared/api/events.js
@@ -49,15 +49,16 @@ export default {
      * @param {...any} context - The context to which the hook applies
      *  (could be for example, a {@link _converse.ChatBox}).
      * @param {...any} data - The data structure to be intercepted and modified by the hook listeners.
+     * @param {...any} extra_args - Additional passthrough arguments for hook listeners.
      * @returns {Promise<any>} - A promise that resolves with the modified data structure.
      */
-    hook (name, context, data) {
+    hook (name, context, data, ...extra_args) {
         const events = _converse._events[name] || [];
         if (events.length) {
             // Create a chain of promises, with each one feeding its output to
             // the next. The first input is a promise with the original data
             // sent to this hook.
-            return events.reduce((o, e) => o.then(d => e.callback(context, d)), Promise.resolve(data));
+            return events.reduce((o, e) => o.then(d => e.callback(context, d, ...extra_args)), Promise.resolve(data));
         } else {
             return data;
         }